:truck: Move tests back to top level folder
