+++ collections/filter_keys_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../../testing/asserts.ts";
-import { filterKeys } from "../filter_keys.ts";
+import { assertEquals } from "../testing/asserts.ts";
+import { filterKeys } from "./filter_keys.ts";
 
 function filterKeysTest<T>(
   input: [Record<string, T>, (key: string) => boolean],
