+++ collections/map_entries_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../../testing/asserts.ts";
-import { mapEntries } from "../map_entries.ts";
+import { assertEquals } from "../testing/asserts.ts";
+import { mapEntries } from "./map_entries.ts";
 
 function mapEntriesTest<T, O>(
   input: [Record<string, T>, (entry: [string, T]) => [string, O]],
