+++ collections/map_keys_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../../testing/asserts.ts";
-import { mapKeys } from "../map_keys.ts";
+import { assertEquals } from "../testing/asserts.ts";
+import { mapKeys } from "./map_keys.ts";
 
 function mapKeysTest<T>(
   input: [Record<string, T>, (key: string) => string],
