+++ collections/map_values_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../../testing/asserts.ts";
-import { mapValues } from "../map_values.ts";
+import { assertEquals } from "../testing/asserts.ts";
+import { mapValues } from "./map_values.ts";
 
 function mapValuesTest<T, O>(
   input: [Record<string, T>, (value: T) => O],
