+++ collections/union_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../../testing/asserts.ts";
-import { union } from "../union.ts";
+import { assertEquals } from "../testing/asserts.ts";
+import { union } from "./union.ts";
 
 function unionTest<I>(
   input: Array<Array<I>>,
